+++ src/main/java/net/geforcemods/securitycraft/blocks/KeypadSmokerBlock.java
@@ -2,7 +2,7 @@ package net.geforcemods.securitycraft.blocks;
 
 import net.geforcemods.securitycraft.SCContent;
 import net.geforcemods.securitycraft.blockentities.AbstractKeypadFurnaceBlockEntity;
-import net.geforcemods.securitycraft.blockentities.KeypadFurnaceBlockEntity;
+import net.geforcemods.securitycraft.blockentities.KeypadSmokerBlockEntity;
 import net.minecraft.core.BlockPos;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.block.entity.BlockEntity;
@@ -10,18 +10,18 @@ import net.minecraft.world.level.block.entity.BlockEntityTicker;
 import net.minecraft.world.level.block.entity.BlockEntityType;
 import net.minecraft.world.level.block.state.BlockState;
 
-public class KeypadFurnaceBlock extends AbstractKeypadFurnaceBlock {
-	public KeypadFurnaceBlock(Properties properties) {
+public class KeypadSmokerBlock extends AbstractKeypadFurnaceBlock {
+	public KeypadSmokerBlock(Properties properties) {
 		super(properties);
 	}
 
 	@Override
 	public <T extends BlockEntity> BlockEntityTicker<T> getTicker(Level level, BlockState state, BlockEntityType<T> type) {
-		return level.isClientSide ? null : createTickerHelper(type, SCContent.beTypeKeypadFurnace, AbstractKeypadFurnaceBlockEntity::serverTick);
+		return level.isClientSide ? null : createTickerHelper(type, SCContent.beTypeKeypadSmoker, AbstractKeypadFurnaceBlockEntity::serverTick);
 	}
 
 	@Override
 	public BlockEntity newBlockEntity(BlockPos pos, BlockState state) {
-		return new KeypadFurnaceBlockEntity(pos, state);
+		return new KeypadSmokerBlockEntity(pos, state);
 	}
 }