+++ Changelog.md
@@ -1,1 +1,7 @@
+--------------------------Changelog for v1.9.1 of SecurityCraft--------------------------
+
+- New: Password-protected Smoker and Password-protected Blast Furnace
+- Fix: Cannot access the recipe book for the Password-protected Furnace
+- Fix: Password-protected Furnace is not shown as a recipe catalyst for furnace recipes in JEI
+
 --------------------------Changelog for v1.9.0.2-beta1 of SecurityCraft--------------------------